+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable changes to this project will be documented in this file.
 - Corrected duplicate detection for custom cache implementations.
 - Fixed panic caused by invalid bounds in the `range` function.
 - Resource config files imported now allow (and ignore) a `tests` field.
+- Fixed an issue where the `aws_kinesis` input would fail to back off during unyielding read attempts.
 
 ## 3.62.0 - 2022-01-21